Add new check to ensure sass builds
This is a check based on the sass build in:
https://github.com/alphagov/govuk-frontend/blob/main/.github/workflows/sass.yaml

We use an up to date version of sass, < 2.x running on the most recent lts node version.
We install all of our dev dependencies, output the sass version, and use our npm command to build (since this is what we are interested in).